% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_measures.R
\name{absolute_error}
\alias{absolute_error}
\title{Compute absolute error}
\usage{
absolute_error(quantile, value, actual_value)
}
\arguments{
\item{quantile}{vector of forecasted quantiles}

\item{value}{vector of forecasted values}

\item{actual_value}{Actual value.}
}
\description{
Computes absolute error between the actual value and the median of the
forecast distribution.
}
