% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{aggregate_cards}
\alias{aggregate_cards}
\title{Aggregate cards from a list into a single unnested data frame}
\usage{
aggregate_cards(list_of_cards)
}
\arguments{
\item{list_of_cards}{List of prediction or evaluation cards.  See documentation for the outputs
of \code{evalcast::get_predictions()} and \code{evalcast::evaluate_predictions()} for the required
format.}
}
\value{
Data frame such that:
\item There is a one-to-one correspondence between rows of the output and the rows of
\code{card$forecast_distribution} across each \code{card} in \code{list_of_cards}.  That is to say, if each
card has 7 rows in its \code{forecast_distribution} and there are 3 such cards in
\code{list_of_cards}, the output will have 21 rows.
\item The columns of the output correspond to:
\itemize{
\item Columns of \code{card$forecast_distribution}
\item Columns of \code{card}
\item Attributes of \code{card}
}
}
\description{
Aggregate cards from a list into a single unnested data frame
}
