% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{all_attr}
\alias{all_attr}
\title{Return list of attributes}
\usage{
all_attr(cards, attribute)
}
\arguments{
\item{cards}{List of predictions cards or a list of score cards.}

\item{attribute}{Name of attribute.}
}
\description{
Given a list of cards, returns a list of the same length giving the values
of that attribute across all cards.
}
