% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_cards.R
\name{as.predictions_cards}
\alias{as.predictions_cards}
\alias{as.predictions_cards.predictions_cards}
\alias{as.predictions_cards.data.frame}
\title{Convert forecasts into a format compatible with all \code{predictions_cards}}
\usage{
as.predictions_cards(x, ...)

\method{as.predictions_cards}{predictions_cards}(x, ...)

\method{as.predictions_cards}{data.frame}(
  x,
  forecaster = "forecaster",
  forecast_date = lubridate::today(),
  incidence_period = c("epiweek", "day"),
  data_source = "data_source",
  signal = "signal",
  target_end_date = NULL
)
}
\arguments{
\item{x}{Object to be converted. See Methods section below for details on
formatting of each input type.}

\item{...}{Additional arguments passed to methods.}

\item{forecaster}{The name of the forecaster that produced this data.}

\item{forecast_date}{Date object or string of the form "YYYY-MM-DD",
indicating date on which forecasts will be made about some period (e.g.,
epiweek). For example, if \code{forecast_date = "2020-05-11"}, \code{incidence_period = "day"}, and \code{ahead = 3}, then, we'd be making forecasts for "2020-05-14".}

\item{incidence_period}{String indicating the incidence period, either
"epiweek" or "day".}

\item{data_source}{The name of the data_source for the forecast target.}

\item{signal}{The name of the signal for the forecast target.}

\item{target_end_date}{The date the forecast target is observed (combines
\code{forecast_date}, \code{ahead}, and \code{incidence_period}). This is determined
automatically if \code{NULL} (the default).}
}
\value{
Long data frame of forecasts with a class of \code{predictions_cards}.
The first 4 columns are the same as those returned by the forecaster. The
remainder specify the prediction task, 10 columns in total:
\code{ahead}, \code{geo_value}, \code{quantile}, \code{value}, \code{forecaster}, \code{forecast_date},
\code{data_source}, \code{signal}, \code{target_end_date}, and \code{incidence_period}. Here
\code{data_source} and \code{signal} correspond to the response variable only.
}
\description{
Several methods are provided to convert external forecasts given as a
data.frame into \code{predictions_cards} objects compatible with the various
evaluation and plotting functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{predictions_cards}: Simply returns the \code{predictions_cards} object
unchanged.

\item \code{data.frame}: The input data frame \code{x} must contain the
columns \code{ahead}, \code{geo_value}, \code{quantile} and \code{value}. Addional columns
\code{forecaster} \code{forecast_date}, \code{incidence_period}, \code{data_source},
\code{signal}, and \code{target_end_date} will be created if missing.
Other columns will be preserved as-is.
}}

