% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek_processing.R
\name{average_to_epiweek}
\alias{average_to_epiweek}
\alias{sum_to_epiweek}
\title{Summarize value over the trailing epiweek}
\usage{
average_to_epiweek(daily_df, return_sum = TRUE)

sum_to_epiweek(daily_df)
}
\arguments{
\item{daily_df}{a data frame containing the columns \code{time_value}, \code{geo_value},
and \code{value}. Additional columns are ignored}

\item{return_sum}{multiply the average by 7 to compute the total cases per
week (the default) or return the average over the week}
}
\value{
a data frame of the same type with 1 row per complete epiweek. The
\code{time_value} entries will be modified to correspond to the last day
(a Saturday) of each epiweek.

These functions vary only in the manner of the calculation. \verb{average_*()}
computes the average over the week ignoring \code{NA}'s and returns the result.
\verb{sum_*} sums over the epiweek. Thus in the absence of missing values, the
results (when \code{return_sum = TRUE}) will be the same. However, if there are
missing values, the result of \verb{sum_*()} will be lower but will accurately
reflect the reported totals.
}
\description{
Summarize value over the trailing epiweek
}
