% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{averaging_checks}
\alias{averaging_checks}
\title{Check if quantile forecasts are commensurate across groups}
\usage{
averaging_checks(
  cards,
  grp_vars = c("forecaster", "forecast_date", "ahead"),
  avg_vars = c("geo_value"),
  return_checks = FALSE
)
}
\arguments{
\item{cards}{a data frame containing at least the columns named
in \code{grp_vars}, \code{avg_vars} and a column called \code{quantile}}

\item{grp_vars}{character vector of named columns in the score_card at which
average performance will be returned}

\item{avg_vars}{character vector of named columns in the score_card over which
averages will be computed}

\item{return_checks}{useful for debugging. returns a data frame indicating
which groups have valid or invalid avg_vars}
}
\value{
filtered cards based on which groups are completely valid
(meaning "correct" quantiles are present for all \code{avg_vars} in the group)
}
\description{
Check if quantile forecasts are commensurate across groups
}
