% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{collapse_cards}
\alias{collapse_cards}
\title{Reduce size of cards}
\usage{
collapse_cards(cards)
}
\arguments{
\item{cards}{either predictions_cards or scorecards}
}
\value{
cards of the same class but with only one row for each
geo_value/forecast_date/ahead/forecaster (the point estimate)
}
\description{
Reduce size of cards
}
