% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{compute_actual_vs_nominal_prob}
\alias{compute_actual_vs_nominal_prob}
\title{Compute calibration of a quantile forecaster}
\usage{
compute_actual_vs_nominal_prob(
  score_card,
  grp_vars = c("forecaster", "forecast_date", "ahead"),
  avg_vars = c("geo_value")
)
}
\arguments{
\item{score_card}{tibble containing at least columns actual, quantile,
value and any grouping or averaging variables named in the next arguments}

\item{grp_vars}{character vector of named columns in the score_card at which
average performance will be returned}

\item{avg_vars}{character vector of named columns in the score_card over which
averaging performance will be computed}
}
\value{
A tibble containing grp_vars, nominal_prob (the claimed interval
coverage), prop_below (the proportion of actual values falling below
the forecast quantile) and prop_above (the proportion of
actual values falling above the forecast quantile). All
proportions are calculated for each available probability at each
combination of grouping variables by averaging over any variables listed
in avg_vars.
}
\description{
Compute calibration of a quantile forecaster
}
\details{
Note that no checks are performed to ensure that averaging variables
all contain the same predicted quantiles. avg_vars and grp_vars must
have an empty intersection.
}
