% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_signals.R
\name{covidcast_signal_wrapper}
\alias{covidcast_signal_wrapper}
\title{The thinnest wrapper possible. For mock testing, see:
https://krlmlr.github.io/mockr/articles/mockr.html#write-wrapper-functions}
\usage{
covidcast_signal_wrapper(...)
}
\arguments{
\item{...}{same arguments as covidcast_signal}
}
\description{
The thinnest wrapper possible. For mock testing, see:
https://krlmlr.github.io/mockr/articles/mockr.html#write-wrapper-functions
}
