% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covidhub_predictions.R
\name{covidhub_probs}
\alias{covidhub_probs}
\title{Vector of quantiles used for submission to the COVID 19 Forecast Hub}
\usage{
covidhub_probs(type = c("standard", "inc_case"))
}
\arguments{
\item{type}{defaults to the "standard" set of 23 quantiles. Setting
\code{type = "inc_case"} will return a set of 7 quantiles used for incident
cases.}
}
\value{
a numeric vector
}
\description{
See the \href{https://github.com/reichlab/covid19-forecast-hub/blob/master/data-processed/README.md#quantile}{Forecast Hub Documentation}
for more details.
}
\examples{
covidhub_probs()
}
