% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_signals.R
\name{download_signal}
\alias{download_signal}
\title{Wrapper around \code{covidcast::covidcast_signal()} with additional logging.}
\usage{
download_signal(...)
}
\arguments{
\item{...}{arguments to be passed to \code{covidcast::covidcast_signal()}.}
}
\value{
\code{covidcast_signal} data frame.
}
\description{
Wrapper around \code{covidcast::covidcast_signal()} with additional logging.
}
