% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_signals.R
\name{download_signals}
\alias{download_signals}
\title{Wrapper around \code{covidcast::covidcast_signals()} with additional logging, error checking, and
aggregation.}
\usage{
download_signals(
  ...,
  signal_aggregation = "long",
  signal_aggregation_dt = NULL
)
}
\arguments{
\item{...}{arguments to be passed to \code{covidcast::covidcast_signals()}}

\item{signal_aggregation}{format of aggregated data ("wide" or "long").
See \code{covidcast::aggregate_signals()}.}

\item{signal_aggregation_dt}{vector of shifts to apply to aggregated data
See \code{covidcast::aggregate_signals()}.}
}
\value{
list of \code{covidcast_signal} data frames
}
\description{
Wrapper around \code{covidcast::covidcast_signals()} with additional logging, error checking, and
aggregation.
}
