% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalcast-package.R
\docType{package}
\name{evalcast-package}
\alias{evalcast-package}
\title{evalcast}
\description{
A package to evaluate forecasters using the \code{covidcast} R package.
}
\author{
Jacob Bien
Maintainer: Jacob Bien \href{mailto:jbien@usc.edu}{jbien@usc.edu}
}
\keyword{package}
