% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_processing_utils.R
\name{filter_predictions}
\alias{filter_predictions}
\title{Drop irrelevant predictions}
\usage{
filter_predictions(predictions, forecast_type, incidence_period, signal)
}
\arguments{
\item{predictions}{Dataframe as produced by \code{process_target}.}

\item{forecast_type}{"quantile", "point", or both}

\item{incidence_period}{"epiweek", "day", or both}

\item{signal}{Any combination of "confirmed_incidence_num",
"deaths_incidence_num", "deaths_cumulative_num", and/or
"confirmed_admissions_covid_1d".}
}
\value{
predictions dataframe with irrelevant rows removed
}
\description{
Drop irrelevant predictions
}
