% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{find_invalid_forecaster_output}
\alias{find_invalid_forecaster_output}
\title{Check if some of a  forecaster's output is invalid}
\usage{
find_invalid_forecaster_output(pred_card)
}
\arguments{
\item{pred_card}{Prediction card, in the form created by \code{\link[=get_predictions]{get_predictions()}}.}
}
\description{
Check if some of a  forecaster's output is invalid
}
