% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.R
\name{format_wedgeplot}
\alias{format_wedgeplot}
\title{ggplot formatting for wedge calibration plot.}
\usage{
format_wedgeplot(calib, grp_vars, facet_rows, facet_cols)
}
\arguments{
\item{calib}{the output from \code{\link[=compute_calibration]{compute_calibration()}}}

\item{grp_vars}{variables over which to compare calibration. These
determines the color of the lines and faceting depending on \code{type}}

\item{facet_rows}{A variable name to facet data over. Creates a
separate row of plots for each value of specified variable. Can be used
with \code{facet_cols} to create a grid of plots.  Should be passed to
plot_calibration when customized.}

\item{facet_cols}{Same as \code{facet_rows}, but with columns.}
}
\value{
ggplot object
}
\description{
ggplot formatting for wedge calibration plot.
}
