% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_period.R
\name{get_target_ahead}
\alias{get_target_ahead}
\title{Compute forecast target date in days}
\usage{
get_target_ahead(forecast_date, incidence_period, ahead)
}
\arguments{
\item{forecast_date}{Date object or string of the form "YYYY-MM-DD",
indicating date on which forecasts will be made about some period (e.g.,
epiweek). For example, if \code{forecast_date = "2020-05-11"}, \code{incidence_period = "day"}, and \code{ahead = 3}, then, we'd be making forecasts for "2020-05-14".}

\item{incidence_period}{String indicating the incidence period, either
"epiweek" or "day".}

\item{ahead}{Vector of (one or more) integers. How many epiweeks/days ahead
are you forecasting? If \code{incidence_period = "epiweek"} and forecast_date is
Sunday or Monday, then \code{ahead = 1} means the epiweek that includes the
forecast date; if \code{forecast_date} falls on a Tuesday through Saturday, then
it is the following epiweek. If \code{incidence_period = "day"}, then \code{ahead = 1} means the day after forecast date.}
}
\description{
This function is most useful for converting a forecast date and an ahead
specification from epiweek to days. It allows days as well.
}
\examples{
fd <- "2021-03-22" # a monday
get_target_ahead(fd, "epiweek", 1) # forecast the following Saturday
get_target_ahead("2021-03-23", "epiweek", 1) # forecast the Saturday after next
get_target_ahead(fd, "day", 1) # forecast tomorrow
}
\seealso{
\code{\link[=get_target_period]{get_target_period()}}
}
