% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{get_target_period}
\alias{get_target_period}
\title{Get the target period for a forecast date, incidence period and ahead}
\usage{
get_target_period(forecast_date, incidence_period, ahead)
}
\arguments{
\item{forecast_date}{the forecast date}

\item{incidence_period}{either "epiweek" or "day"}

\item{ahead}{vector of (one or more) integers.  How many epiweeks/days ahead
are you forecasting?  If incidence_period is "epiweek" and forecast_date is
Sunday or Monday, then ahead = 1 means the epiweek that includes the
forecast date;  if forecast_date is Tuesday-Saturday, then it is the
following epiweek. If incidence_period is "day" then ahead = 1 means the
day after forecast date.}
}
\description{
Get the target period for a forecast date, incidence period and ahead
}
