% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{intersect_locations}
\alias{intersect_locations}
\title{Remove locations that are not in all cards}
\usage{
intersect_locations(cards)
}
\arguments{
\item{cards}{a list of predictions cards or a list of scorecards}
}
\description{
Remove locations that are not in all cards
}
