% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_measures.R
\name{interval_coverage}
\alias{interval_coverage}
\title{Generate interval coverage error measure function}
\usage{
interval_coverage(coverage)
}
\arguments{
\item{coverage}{Nominal interval coverage (from 0 to 1).}
}
\description{
Returns an error measure function indicating whether a central interval
covers the actual value. The interval is defined as the (alpha/2)-quantile
to the (1 - alpha/2)-quantile, where alpha = 1 - coverage.
}
