% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_processing_utils.R
\name{location_2_geo_value}
\alias{location_2_geo_value}
\title{Convert location FIPS to abbreviations as needed for the pcard format}
\usage{
location_2_geo_value(predictions, default_process_fn = as.character)
}
\arguments{
\item{predictions}{Dataframe with \code{location} column.}

\item{default_process_fn}{Function to use to reformat the \code{location} column
if already in abbreviation format. Defaults to \code{as.character}.}
}
\value{
predictions dataframe with irrelevant columns removed
}
\description{
Convert location FIPS to abbreviations as needed for the pcard format
}
