% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Plot the calibration curves for a scorecard}
\usage{
plot_calibration(scorecard, type = c("wedgeplot", "traditional"), alpha = 0.1)
}
\arguments{
\item{scorecard}{a scorecard tibble}

\item{type}{one of \code{"wedgeplot"} or \code{"traditional"}.}

\item{alpha}{the alpha value}
}
\value{
a ggplot object
}
\description{
Plot the calibration curves for a scorecard
}
