% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_canonical.R
\name{plot_canonical}
\alias{plot_canonical}
\title{A canonical function for creating plots within evalcast}
\usage{
plot_canonical(
  df,
  x,
  y,
  facet_rows = NULL,
  facet_cols = NULL,
  grp_vars = "forecaster",
  aggr = mean,
  dots = TRUE,
  lines = TRUE,
  base_forecaster = NULL,
  scale_before_aggr = FALSE
)
}
\arguments{
\item{df}{dataframe containing the data to plot}

\item{x}{Name of the variable in \code{df} to be used for the x-axis}

\item{y}{name of the variable in \code{df} to be used for the y-axis}

\item{facet_rows}{A variable name to facet data over. Creates a
separate row of plots for each value of specified variable. Can be used
with \code{facet_cols} to create a grid of plots.}

\item{facet_cols}{Same as \code{facet_rows}, but with columns.}

\item{grp_vars}{The set of variable names in \code{df} which to aggregate over.
"forecaster" by default.}

\item{aggr}{function to use to aggregate data over \code{grp_vars}}

\item{dots}{when set to TRUE, a dot layer is included in the plot}

\item{lines}{when set to TRUE, a line layer is included in the plot}

\item{base_forecaster}{If set, scales the y-value for all forecasters by
the corresponding values of the \code{base_forecaster}}

\item{scale_before_aggr}{If TRUE, scales results by the
\code{base_forecaster} before aggregating by the \code{aggr} function. FALSE by
default.}
}
\description{
A canonical function for creating plots within evalcast
}
