% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot the interval coverage}
\usage{
plot_coverage(scorecards, alpha = 0.2, type = c("all", "one"))
}
\arguments{
\item{scorecards}{a list of different forecasters scorecards, all on the
same forecasting task (i.e., same ahead, etc.)}

\item{location}{of vertical line}
}
\description{
Plot the interval coverage
}
