% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_measure.R
\name{plot_measure}
\alias{plot_measure}
\title{Plot error measure}
\usage{
plot_measure(scorecards, err_name, type = "boxplot")
}
\arguments{
\item{scorecards}{List of different score cards, all on the same forecasting
task (i.e., same ahead, etc.).}

\item{err_name}{Name of a column appearing in all score cards.}

\item{type}{One of "boxplot" or "dotplot".}
}
\description{
Plot error measure
}
