% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot predictions along with truth}
\usage{
plot_trajectory(
  predictions_cards,
  intervals = c(0.5, 0.8, 0.95),
  plot_it = TRUE,
  show_geo_value = NULL,
  show_forecaster = NULL,
  side_truth = NULL,
  show_points = TRUE,
  show_quantiles = TRUE,
  geo_type = c("county", "hrr", "msa", "dma", "state", "hhs", "nation"),
  ...
)
}
\arguments{
\item{predictions_cards}{long data frame of forecasts with a class of
\code{predictions_cards} as created by \code{\link[=get_predictions]{get_predictions()}} or
\code{\link[=get_zoltar_predictions]{get_zoltar_predictions()}} or \code{\link[=get_covidhub_predictions]{get_covidhub_predictions()}}.
The first 4 columns are the same as those returned by the forecaster. The
remainder specify the prediction task, 10 columns in total:
\code{ahead}, \code{geo_value}, \code{quantile}, \code{value}, \code{forecaster}, \code{forecast_date},
\code{data_source}, \code{signal}, \code{target_end_date}, and \code{incidence_period}. Here
\code{data_source} and \code{signal} correspond to the response variable only.}

\item{intervals}{vector of confidence intervals to show. More than 3 is ugly
and will be reduced to the default set.}

\item{plot_it}{should we actually produce the figure. If you have many
geo_values or forecasters, best not to. Then you can assign the output
to an object and facet or apply aesthetic specifications to meet your
own standards.}

\item{show_geo_value}{a subset of \code{geo_values} in your \code{predictions_cards}.
Default uses all available. If you only want to see some, this may speed
up the call.}

\item{show_forecaster}{a subset of \code{forecasters} in your \code{predictions_cards}.
Default uses all available. If you only want to see some, this may speed
up the call.}

\item{side_truth}{optional. Use your own truth data rather than downloading
with \code{\link[covidcast:covidcast_signal]{covidcast::covidcast_signal()}}. Must contain columns \code{geo_value},
\code{value} and \code{target_end_date}}

\item{show_points}{do you want points as well as lines}

\item{show_quantiles}{do you want to plot the quantiles or just lines}

\item{geo_type}{String indicating geographical type, such as "county", or
"state". See the \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_geography.html}{COVIDcast Geographic Coding documentation}
for available options.}

\item{...}{additional arguments passed to \code{\link[covidcast:covidcast_signal]{covidcast::covidcast_signal()}}}
}
\value{
invisibly returns a ggplot object
}
\description{
Plot predictions along with truth
}
