% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_signals.R
\name{populate_cache}
\alias{populate_cache}
\title{A function to download and cache signals from the Covidcast API for future use. It is recommended that you use
this function to download more data than you expect than you will need, since the local data cache can be easily
filtered, but updating a local cache with more remote values requires complicated logic we have not implemented.}
\usage{
populate_cache(source_signals, offline_signal_dir)
}
\arguments{
\item{source_signals}{a tibble with the following columns: \code{data_source}, \code{signal}, \code{start_day}, \code{as_of},}

\item{offline_signal_dir}{the directory that stores the cached data for each
(signal, forecast day) pair.}
}
\value{
no returns.
}
\description{
A function to download and cache signals from the Covidcast API for future use. It is recommended that you use
this function to download more data than you expect than you will need, since the local data cache can be easily
filtered, but updating a local cache with more remote values requires complicated logic we have not implemented.
}
