% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{print.prediction_card}
\alias{print.prediction_card}
\title{Print a single prediction card.}
\usage{
\method{print}{prediction_card}(x, ...)
}
\arguments{
\item{x}{Prediction card.}

\item{...}{Additional arguments to be passed to \code{print.tbl_df}}
}
\description{
Print a single prediction card.
}
