% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{print.predictions_cards}
\alias{print.predictions_cards}
\title{Print method for predictions_cards}
\usage{
\method{print}{predictions_cards}(x, ...)
}
\arguments{
\item{x}{Prediction card.}
}
\description{
Print method for predictions_cards
}
