% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{print.score_card}
\alias{print.score_card}
\title{Print a single score card.}
\usage{
\method{print}{score_card}(x, ...)
}
\arguments{
\item{x}{Score card.}

\item{...}{Additional arguments to be passed to \code{print.tbl_df}}
}
\description{
Print a single score card.
}
