% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{print.score_cards}
\alias{print.score_cards}
\title{Print a single score card.}
\usage{
\method{print}{score_cards}(x, ...)
}
\arguments{
\item{card}{Score card.}
}
\description{
Print a single score card.
}
