% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_processing_utils.R
\name{process_target}
\alias{process_target}
\title{Separate fields of target col and create derived fields}
\usage{
process_target(predictions, remove)
}
\arguments{
\item{predictions}{Dataframe with \code{target} column containing information
about time period and signal forecast.}

\item{remove}{If TRUE, remove input \code{target} column from output data frame.
See \code{dplyr::separate}.}
}
\value{
predictions dataframe with new \code{ahead}, \code{incidence_period}, \code{inc},
\code{response}, \code{data_source}, and \code{signal} fields
}
\description{
This function supports only "confirmed_incidence_num",
"deaths_incidence_num", "deaths_cumulative_num", and/or
"confirmed_admissions_covid_1d".
}
