% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_by_forecaster.R
\name{scale_by_forecaster}
\alias{scale_by_forecaster}
\title{Scale error measures based on those of a particular forecaster.}
\usage{
scale_by_forecaster(
  score_card,
  score_cols,
  base_forecaster_name,
  id_cols = c("forecaster", "ahead", "geo_value", "forecast_date", "data_source",
    "signal", "target_end_date", "incidence_period"),
  drop_base_entries = TRUE
)
}
\arguments{
\item{score_card}{score_card like that returned by \code{evaluate_predictions()}}

\item{score_cols}{vector of column names in \code{score_card} to normalize}

\item{base_forecaster_name}{name of forecaster in \code{score_card$forecaster} column by whose error
values the remaining forecasters' errors will be scaled}

\item{id_cols}{vector of column names in \code{score_card} that identify distinct forecasts (i.e. the
independent variables of \code{score_card}).}

\item{drop_base_entries}{whether to drop the entries in \code{score_card} from
\code{base_forecaster_name} (when FALSE their corresponding \code{score_cols} values will all be 1)}
}
\value{
A tibble whose columns are \code{c(id_cols, score_cols)} whose \code{id_cols} values are copied
directly from \code{score_card} and whose \code{score_cols} values are normalized with respect to
\code{base_forecaster_name}.
}
\description{
Scale error measures based on those of a particular forecaster.
}
