% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures.R
\name{score_func_param_checker}
\alias{score_func_param_checker}
\title{Common parameter checks for score functions}
\usage{
score_func_param_checker(quantiles, values, actual_value, id = "")
}
\arguments{
\item{quantiles}{vector of forecasted quantiles}

\item{values}{vector of forecasted values}

\item{actual_value}{actual_value, either as a scalar or a vector of the same
length as \code{quantiles} and \code{values}}

\item{id}{string to identify the caller of the function and displayed in
error messages (recommended to be the parent function's name)}
}
\description{
A set of common checks for score functions, meant to identify common causes
of issues. Avoids \code{assert_that} for speed.
}
