% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_processing_utils.R
\name{select_pcard_cols}
\alias{select_pcard_cols}
\title{Select and reorder desired columns for the pcard format}
\usage{
select_pcard_cols(predictions)
}
\arguments{
\item{predictions}{Dataframe as produced by \code{filter_predictions}.}
}
\value{
predictions dataframe with irrelevant columns removed
}
\description{
Select and reorder desired columns for the pcard format
}
