% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.R
\name{setup_wedgeplot}
\alias{setup_wedgeplot}
\title{Preprocessing for wedge calibration plot}
\usage{
setup_wedgeplot(calib)
}
\arguments{
\item{calib}{the output from \code{\link[=compute_calibration]{compute_calibration()}}}
}
\value{
a data frame
}
\description{
Preprocessing for wedge calibration plot
}
