% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{unpack_single_card.evaluation_card}
\alias{unpack_single_card.evaluation_card}
\title{Unpack a single evaluation card into an unnested tibble}
\usage{
\method{unpack_single_card}{evaluation_card}(card)
}
\description{
Unpack a single evaluation card into an unnested tibble
}
