% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{unpack_single_card.score_card}
\alias{unpack_single_card.score_card}
\title{Unpack a single score card into an unnested tibble}
\usage{
\method{unpack_single_card}{score_card}(card)
}
\description{
Unpack a single score card into an unnested tibble
}
