% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_measures.R
\name{weighted_interval_score}
\alias{weighted_interval_score}
\title{Compute weighted interval score}
\usage{
weighted_interval_score(quantile_forecasts, actual_value)
}
\description{
For more details, see https://arxiv.org/abs/2005.12881
}
\details{
Bracher, J., Ray, E. L., Gneiting, T., & Reich, N. G. (2020). Evaluating
epidemic forecasts in an interval format. arXiv preprint arXiv:2005.12881.
}
