% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures.R
\name{absolute_error}
\alias{absolute_error}
\title{Compute absolute error}
\usage{
absolute_error(quantile, value, actual_value)
}
\arguments{
\item{quantile}{vector of forecasted quantiles}

\item{value}{vector of forecasted values}

\item{actual_value}{vector of actual values of the same length as
\code{quantile}/\code{value} or a scalar}
}
\description{
Absolute error of a forecaster
}
\details{
Intended to be used with \code{evaluate_predictions()}, it expects three arguments
of the same length, finds the location of the point forecast, and returns
the absolute error.
}
