% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline_forecaster}
\alias{baseline_forecaster}
\title{Baseline forecaster}
\usage{
baseline_forecaster(
  df,
  forecast_date,
  signals,
  incidence_period,
  ahead,
  geo_type
)
}
\arguments{
\item{df}{a data frame of the format that is outputted by
\code{\link[covidcast]{covidcast_signal}}.}

\item{forecast_date}{date on which forecasts will be made about some period
(e.g., epiweek).  For example, if forecast_date is ymd("2020-05-11"),
incidence_period is "day",  and ahead = 3, then, we'd be making forecasts
for "2020-05-14".}

\item{signals}{a tibble with columns "data_source" and "signal" that
specifies which variables from the covidcast API will be used by
my_forecaster. The first row of signals is taken to be the response.   If
using incidence_period "epiweek", the response should be something for
which summing daily values over an epiweek makes sense (e.g., counts or
proportions but not log(counts) or log(proportions)).  Available data
sources and signals are documented in the \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html}{COVIDcast signal documentation}
documentation.  Can optionally include a column "first_day" giving the
earliest day of data needed from that data source.}

\item{incidence_period}{either "epiweek" or "day"}

\item{ahead}{vector of (one or more) integers.  How many epiweeks/days ahead
are you forecasting?  If incidence_period is "epiweek" and forecast_date is
Sunday or Monday, then ahead = 1 means the epiweek that includes the
forecast date;  if forecast_date is Tuesday-Saturday, then it is the
following epiweek. If incidence_period is "day" then ahead = 1 means the
day after forecast date.}

\item{geo_type}{one of the geographical types, matching the terms used on the
api ("county", "dma", "hrr", "msa", "state")}
}
\value{
A data frame with columns "ahead", "location", "probs", "quantiles".
The quantiles column gives the probs-quantile of the forecast distribution
for that location and ahead.
}
\description{
This serves as a template for a forecaster.  It's not intended to be a great
forecaster.
}
