% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{compute_coverage}
\alias{compute_coverage}
\title{Compute observed coverage from a quantile forecaster}
\usage{
compute_coverage(
  predictions_cards,
  geo_type = c("county", "hrr", "msa", "dma", "state", "hhs", "nation"),
  grp_vars = c("forecaster", "forecast_date", "ahead"),
  avg_vars = c("geo_value"),
  backfill_buffer = 0
)
}
\arguments{
\item{grp_vars}{character vector of named columns in the score_card at which
average performance will be returned}

\item{avg_vars}{character vector of named columns in the score_card over which
averaging performance will be computed}
}
\value{
Long data frame of forecasts with a class of \code{predictions_cards}.
The first 4 columns are the same as those returned by the forecaster. The
remainder specify the prediction task, 10 columns in total:
\code{ahead}, \code{geo_value}, \code{quantile}, \code{value}, \code{forecaster}, \code{forecast_date},
\code{data_source}, \code{signal}, \code{target_end_date}, and \code{incidence_period}. Here
\code{data_source} and \code{signal} correspond to the response variable only.

A tibble containing grp_vars, nominal_prob (the claimed interval
coverage), prop_below (the proportion of actual values falling below
the lower end of the confidence interval), prop_above (the proportion of
actual values falling above the upper end of the confidence interval),
and prop_covered (the proportion falling inside the interval). All
proportions are calculated for each available symmetric interval at each
combination of grouping variables by averaging over any variables listed
in avg_vars.
}
\description{
Compute observed coverage from a quantile forecaster
}
\details{
Checks \strong{are} performed to ensure that averaging variables
all contain the same confidence intervals though these may vary over
grouping variables.
avg_vars and grp_vars must
have an empty intersection.
}
