% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_signals.R
\name{download_signal}
\alias{download_signal}
\title{Download signal from covidcast}
\usage{
download_signal(...)
}
\arguments{
\item{...}{Arguments that are passed to \code{\link[covidcast:covidcast_signal]{covidcast::covidcast_signal()}}.}
}
\description{
This is a simple wrapper to \code{\link[covidcast:covidcast_signal]{covidcast::covidcast_signal()}} that is less verbose.
}
