% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_predictions}
\alias{evaluate_predictions}
\title{Create a list of score cards}
\usage{
evaluate_predictions(
  predictions_cards,
  err_measures = list(wis = weighted_interval_score, ae = absolute_error, coverage_80 =
    interval_coverage(alpha = 0.2)),
  backfill_buffer = 10
)
}
\arguments{
\item{predictions_cards}{tibble of predictions
that are all for the same prediction task, meaning they are for the same
response, incidence period, ahead, and geo type. Forecasts may be for a
different forecast date or forecaster.
A predictions card may be created by the function
\code{\link[=get_predictions]{get_predictions()}}, downloaded with \code{\link[=get_covidhub_predictions]{get_covidhub_predictions()}} or
possibly created manually.}

\item{err_measures}{Named list of one or more functions, where each function
takes a data frame with three columns \code{quantile}, \code{value} and \code{actual}
(i.e., observed) returns a scalar measure of error.}

\item{backfill_buffer}{How many days until response is deemed trustworthy
enough to be taken as correct? See details for more.}
}
\value{
tibble of "score cards" with additional columns for each err_measure
}
\description{
Performs backtesting, through the following steps:
\enumerate{
\item Takes a list of predictions cards (as created by
\code{\link[=get_predictions]{get_predictions()}}). These should be from a single forecaster, each card
corresponding to a different forecast date.
\item Downloads from the COVIDcast API the latest available data to compute
what actually occurred (summing the response over the incidence period).
\item Computes various user-specified error measures.
}
The result is a list of "score cards", where each list element corresponds to
a distinct ahead value. A score card is a data frame in which each row
corresponds to a location-day pair, and the columns give the values of the
error measures (along with other information including the forecast
distributions and the actual response values).
}
\details{
Backfill refers to the process by which some data sources go back in time
updating previously reported values. Suppose it is September 14 and we are
evaluating our predictions for what happened in the previous epiweek
(September 6 through 12). Although we may be able to calculate a value for
"actual", we might not trust this value since on September 16, backfill may
occur changing what is known about the period September 6 through 12.  There are
two consequences of this phenomenon.  First, running this function on
different dates may result in different estimates of the error. Second, we
may not trust the evaluations we get that are too recent.  The parameter
\code{backfill_buffer} specifies how long of a buffer period we should
enforce.  This will be dependent on the data source and signal and is left
to the user to determine.  If backfill is not relevant for the particular
signal you are predicting, then you can set \code{backfill_buffer} to 0.
}
