% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_predictions.R
\name{evaluate_predictions}
\alias{evaluate_predictions}
\title{Create a score card data frame}
\usage{
evaluate_predictions(
  predictions_cards,
  truth,
  err_measures = list(wis = weighted_interval_score, ae = absolute_error, coverage_80 =
    interval_coverage(coverage = 0.8)),
  grp_vars = c("forecaster", intersect(colnames(predictions_cards), colnames(truth)))
)
}
\arguments{
\item{predictions_cards}{tibble of quantile forecasts, which contains at
least \code{quantile} and \code{value} columns, as well as any other prediction task
identifiers. For covid data, a predictions card may be created by the
function \code{\link[=get_predictions]{get_predictions()}}, downloaded with \code{\link[=get_covidhub_predictions]{get_covidhub_predictions()}}
or created manually.}

\item{truth}{truth data (observed). This should be a data frame that will be
joined to \code{predictions_cards} by all available columns. The observed data
column should be named \code{actual}.}

\item{err_measures}{Named list of one or more functions, where each function
takes a data frame with three columns \code{quantile}, \code{value} and \code{actual}
(i.e., observed) and returns a scalar measure of error. Null or an empty
list may be provided if scoring is not desired.}

\item{grp_vars}{character vector of named columns in \code{predictions_cards}
such that the combination gives a unique (quantile) prediction.}
}
\value{
tibble of "score cards". Contains the same information as the
\code{predictions_cards()} with additional columns for each \code{err_measure} and
for the truth (named \code{actual}).
}
\description{
Performs testing, through the following steps:
\enumerate{
\item Takes a prediction card (as created by
\code{\link[=get_predictions]{get_predictions()}}).
\item Computes various user-specified error measures.
}
The result is a "score card" data frame, where each row corresponds to
a prediction-result pair, where the columns define the prediction
task, give the observed value, and give the calculated values of the provided
error measures.
}
