% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_predictions_single_ahead}
\alias{evaluate_predictions_single_ahead}
\title{Create score cards for a single ahead value.}
\usage{
evaluate_predictions_single_ahead(
  predictions_cards,
  err_measures,
  backfill_buffer
)
}
\description{
Create score cards for a single ahead value.
}
