% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predictions.R
\name{filter_predictions}
\alias{filter_predictions}
\title{Filter a list of predictions cards based on attributes}
\usage{
filter_predictions(
  predictions_cards,
  name_of_forecaster = NULL,
  response_data_source = NULL,
  response_signal = NULL,
  forecast_date = NULL,
  incidence_period = NULL,
  ahead = NULL,
  geo_type = NULL
)
}
\arguments{
\item{predictions_cards}{a list of predictions cards}
}
\description{
Filter a list of predictions cards based on attributes
}
