% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predictions.R
\name{filter_predictions}
\alias{filter_predictions}
\title{Filter a list of predictions cards based on attributes}
\usage{
filter_predictions(
  predictions_cards,
  name_of_forecaster = NULL,
  response_data_source = NULL,
  response_signal = NULL,
  forecast_date = NULL,
  incidence_period = NULL,
  ahead = NULL,
  geo_type = NULL
)
}
\arguments{
\item{predictions_cards}{a list of predictions cards}

\item{name_of_forecaster}{the name of the forecaster}

\item{response_data_source}{the response data source}

\item{response_signal}{the response signal}

\item{forecast_date}{the forecast date}

\item{incidence_period}{either "epiweek" or "day"}

\item{ahead}{vector of (one or more) integers.  How many epiweeks/days ahead
are you forecasting?  If incidence_period is "epiweek" and forecast_date is
Sunday or Monday, then ahead = 1 means the epiweek that includes the
forecast date;  if forecast_date is Tuesday-Saturday, then it is the
following epiweek. If incidence_period is "day" then ahead = 1 means the
day after forecast date.}

\item{geo_type}{one of the geographical types, matching the terms used on the
api ("county", "dma", "hrr", "msa", "state")}
}
\description{
Filter a list of predictions cards based on attributes
}
