% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-predictions.R
\name{get_actuals}
\alias{get_actuals}
\title{Retrieve actual results for provided forecasts}
\usage{
get_actuals(
  predictions_cards,
  geo_type = c("county", "hrr", "msa", "dma", "state", "hhs", "nation")
)
}
\arguments{
\item{predictions_cards}{tibble of predictions
that are all for the same prediction task, meaning they are for the same
response, incidence period and geo_type. Forecasts may be for a
different forecast date or forecaster.
A predictions card may be created by the function
\code{\link[=get_predictions]{get_predictions()}}, downloaded with \code{\link[=get_covidhub_predictions]{get_covidhub_predictions()}} or
possibly created manually.}

\item{geo_type}{String indicating geographical type, such as "county", or
"state". See the \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_geography.html}{COVIDcast Geographic Coding documentation}
for available options.}
}
\value{
'predictions_cards' with an added column \code{actual}, which represents
the observed value on the date. The \code{quantile} and \code{value} columns are
dropped, as the actual value does not depend on the quantile predictions.
}
\description{
Retrieve actual results for provided forecasts
}
