% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covidhub_predictions.R
\name{get_covidhub_forecaster_names}
\alias{get_covidhub_forecaster_names}
\title{List all COVID forecast models available}
\usage{
get_covidhub_forecaster_names(repo = c("zoltar", "covid19forecast_repo"))
}
\arguments{
\item{repo}{character strinng either "zoltar" indicating the
\href{https://zoltardata.com}{Zoltar} Forecast Archive or "covid19forecast_repo"
which lists those available at the
\href{https://github.com/reichlab/covid19-forecast-hub}{Reichlab}
Github submission repo.}
}
\value{
character vector of all available forecasters
}
\description{
Utility function to list all forecasters submitting COVID-19 forecasts to
the \href{http://covid19forecasthub.org/}{COVID 19 Forecast Hub}.
}
