% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Get predictions}
\usage{
get_predictions(
  forecaster,
  name_of_forecaster,
  signals,
  forecast_dates,
  incidence_period,
  ahead,
  geo_type,
  geo_values = "*",
  apply_corrections = NULL
)
}
\arguments{
\item{forecaster}{a function that outputs a tibble with columns...}

\item{name_of_forecaster}{the name of forecaster}

\item{signals}{a tibble with columns "data_source" and "signal" that
specifies which variables from the covidcast API will be used by
my_forecaster. The first row of signals is taken to be the response.   If
using incidence_period "epiweek", the response should be something for
which summing daily values over an epiweek makes sense (e.g., counts or
proportions but not log(counts) or log(proportions)).  Available data
sources and signals are documented in the \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html}{COVIDcast signal documentation}
documentation.  Can optionally include a column "first_day" giving the
earliest day of data needed from that data source.}

\item{forecast_dates}{a vector of class Date on which forecasts will be made.
e.g. \code{c(lubridate::ymd("2020-09-07"), lubridate::ymd("2020-09-14"))}}

\item{incidence_period}{either "epiweek" or "day"}

\item{ahead}{vector of (one or more) integers.  How many epiweeks/days ahead
are you forecasting?  If incidence_period is "epiweek" and forecast_date is
Sunday or Monday, then ahead = 1 means the epiweek that includes the
forecast date;  if forecast_date is Tuesday-Saturday, then it is the
following epiweek. If incidence_period is "day" then ahead = 1 means the
day after forecast date.}

\item{geo_type}{one of the geographical types, matching the terms used on the
api ("county", "dma", "hrr", "msa", "state")}

\item{geo_values}{see \link[covidcast]{covidcast_signal} for a description
of this parameter.}

\item{apply_corrections}{an optional function that applies data corrections
to the signals. Input is a data frame as returned by \code{download_signals()}.
The returned object must be the same but with additional variables.
Corrected values must exist in the column \code{corrected}.}
}
\value{
a list of predictions cards
}
\description{
For each of the provided forecast dates, runs a forecaster using the data
that would have been available as of that given forecast date. Returns a list
of "predictions cards."  A prediction card is a data frame giving the
forecast distributions of a given forecaster for a given forecast task.  A
forecast task is specified by the forecast date, ahead, response, incidence
period, and geo_type (e.g., 1-epiweek ahead death forecasting at the state
level with predictions made using the information as of Sept. 14).
}
\details{
A predictions card has two columns:
\enumerate{
\item location - the FIPS code of the location.  For counties, this is the
same as the geo_value used in the API.  However, for states, the location
and geo_value are different because the API uses state abbreviations instead
of FIPS for geo_value.
\item forecast_distribution - this is a list column... each element itself
contains a tibble with the covidhub quantiles.}
A predictions card has attributes that specify the exact forecasting
task that was being carried out and the name of the forecaster.
}
\examples{
baby_predictions = get_predictions(
  baseline, "baby",
  tibble(
    data_source=c("jhu-csse", "usa-facts"),
    signal = c("deaths_incidence_num","confirmed_incidence_num"),
    start_day=lubridate::ymd("2020-09-15")),
  lubridate::ymd("2020-10-01"),"epiweek", 1L, "state", "mi")

baby_correct <- function(x) x \%>\% dplyr::mutate(corrected = 2*value)

baby_corrected = get_predictions(
  baseline, "baby",
  tibble(
    data_source=c("jhu-csse", "usa-facts"),
    signal = c("deaths_incidence_num","confirmed_incidence_num"),
    start_day=lubridate::ymd("2020-09-15")),
  lubridate::ymd("2020-10-01"),"epiweek", 1L, "state", "mi",
  apply_corrections = baby_correct)

}
