% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Get predictions}
\usage{
get_predictions(
  forecaster,
  name_of_forecaster,
  signals,
  forecast_dates,
  incidence_period,
  ahead,
  geo_type,
  apply_corrections = NULL,
  signal_aggregation = c("list", "wide", "long"),
  signal_aggregation_dt = NULL,
  as_of_override = function(forecast_date) forecast_date,
  ...
)
}
\arguments{
\item{forecaster}{Function that outputs a tibble with columns \code{ahead},
\code{geo_value}, \code{quantile}, and \code{value}. The \code{quantile} column gives the
probabilities associated with quantile forecasts for that location and
ahead. If your forecaster produces point forecasts, then set \code{quantile=NA}.}

\item{name_of_forecaster}{String indicating name of the forecaster.}

\item{signals}{Tibble with columns \code{data_source} and \code{signal} that specifies
which variables from the COVIDcast API will be used by \code{forecaster}. Each
row of \code{signals} represents a separate signal, and first row is taken to be
the response. If using \code{incidence_period = "epiweek"}, the response should
be something for which summing daily values over an epiweek makes sense
(e.g., counts or proportions but not log(counts) or log(proportions)).
Available data sources and signals are documented in the \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html}{COVIDcast signal documentation}.
An optional column \code{start_day} can also be included. This can be a Date
object or string in the form "YYYY-MM-DD", indicating the earliest date of
data needed from that data source. Importantly, \code{start_day} can also be a
function (represented as a list column) that takes a forecast date and
returns a start date for model training (again, Date object or string in
the form "YYYY-MM-DD"). The latter is useful when the start date should be
computed dynamically from the forecast date (e.g., when \code{forecaster} only
trains on the most recent 4 weeks of data).}

\item{forecast_dates}{Vector of Date objects (or strings of the form
"YYYY-MM-DD") indicating dates on which forecasts will be made.}

\item{incidence_period}{String indicating the incidence period, either
"epiweek" or "day".}

\item{ahead}{Vector of (one or more) integers. How many epiweeks/days ahead
are you forecasting? If \code{incidence_period = "epiweek"} and forecast_date is
Sunday or Monday, then \code{ahead = 1} means the epiweek that includes the
forecast date; if \code{forecast_date} falls on a Tuesday through Saturday, then
it is the following epiweek. If \code{incidence_period = "day"}, then \code{ahead = 1} means the day after forecast date.}

\item{geo_type}{String indicating geographical type, such as "county", or
"state". See the \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_geography.html}{COVIDcast Geographic Coding documentation}
for available options.}

\item{apply_corrections}{an optional function that applies data corrections
to the signals. Input is a data frame or list as returned as
\code{df <- covidcast::download_signals()}.
The returned object should be of the type expected by your forecaster.
This function will be called as \code{apply_corrections(df)}.}

\item{signal_aggregation}{this and the next argument control the type of
data your forecaster expects to receive from covidcast. By default,
different signals are passed in "list" format. But you may alternatively
request "wide" or "long". See \code{\link[covidcast:covidcast_signals]{covidcast::covidcast_signals()}} and
\code{\link[covidcast:aggregate_signals]{covidcast::aggregate_signals()}} for more details.}

\item{signal_aggregation_dt}{for any data format,
\code{\link[covidcast:aggregate_signals]{covidcast::aggregate_signals()}} can perform leading and lagging for you.
See that documentation for more details.}

\item{as_of_override}{by default, the \code{as_of} date of data downloaded from
covidcast is loaded with \code{as_of = forecast_date}. This means that data
is "rewound" to days in the past. Any data revisions made since, would
not have been present at that time, and would not be available to the
forecaster. It's likely, for example, that no data would actually exist
for the forecast date on the forecast date (there is some latency between
the time signals are reported and the dates for which they are reported).
You can override this functionality, though we strongly advise you do so
with care, by passing a function of a single forecast_date here. The
function should return a date.}

\item{...}{Additional named arguments to be passed to \code{forecaster()}}

\item{geo_values}{Vector or list of locations whose data will be passed to
\code{forecaster}. The default, "*", fetches all that is available in the API.}
}
\value{
Long data frame of forecasts with a class of \code{predictions_cards}.
The first 4 columns are the same as those returned by the forecaster. The
remainder specify the prediction task, 10 columns in total:
\code{ahead}, \code{geo_value}, \code{quantile}, \code{value}, \code{forecaster}, \code{forecast_date},
\code{data_source}, \code{signal}, \code{target_end_date}, and \code{incidence_period}. Here
\code{data_source} and \code{signal} correspond to the response variable only.
}
\description{
For each of the provided forecast dates, runs a forecaster using the data
that would have been available as of that given forecast date. Returns a list
of "predictions cards", where each list element corresponds to a different
forecast date. A predictions card is a data frame giving the forecast
distributions of a given forecaster for a given forecast task. A forecast
task is specified by the forecast date, ahead, response, incidence period,
and geo type (e.g., 1-epiweek-ahead death forecasting at the state level with
predictions made using the information as of September 14).
}
\examples{
\dontrun{
baby_predictions = get_predictions(
  baseline_forecaster, "baby",
  tibble::tibble(
    data_source=c("jhu-csse", "usa-facts"),
    signal = c("deaths_incidence_num","confirmed_incidence_num"),
    start_day="2020-08-15"), 
  "2020-10-01","epiweek", 1:4, 
  "state", "mi", signal_aggregation="long")
}

}
