% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Get predictions}
\usage{
get_predictions(
  forecaster,
  name_of_forecaster,
  signals,
  forecast_dates,
  incidence_period = c("epiweek", "day"),
  apply_corrections = function(signals) signals,
  response_data_source = signals$data_source[1],
  response_data_signal = signals$signal[1],
  forecaster_args = list(),
  parallel_execution = TRUE,
  additional_mclapply_args = list(),
  honest_as_of = TRUE,
  offline_signal_dir = NULL
)
}
\arguments{
\item{forecaster}{Function that outputs a tibble with columns \code{ahead},
\code{geo_value}, \code{quantile}, and \code{value}. The \code{quantile} column gives the
probabilities associated with quantile forecasts for that location and
ahead. If your forecaster produces point forecasts, then set \code{quantile=NA}.

One argument to \code{forecaster} must be named \code{df_list}. It will be
populated with the list of historical data returned by a call
to COVIDcast. The list will be the same length as the number of rows in
the \code{signals} tibble (see below).
The forecaster will also receive a single \code{forecast_date} as a named argument.
Any additional named arguments can be passed via the \code{forecaster_args}
argument below.

Thus, the forecaster should have a signature like
\code{forecaster(df_list = data, forecast_data = forecast_date, ...)}}

\item{name_of_forecaster}{String indicating name of the forecaster.}

\item{signals}{Tibble with mandatory columns \code{data_source} and \code{signal} and
optional columns \code{start_day}, \code{as_of}, \code{geo_typ}, \code{geo_values}.

\code{data_source} and \code{signal} specify which variables from the COVIDcast API
will be used by \code{forecaster}. Each
row of \code{signals} represents a separate signal, and first row is taken to be
the response unless explicitly overridden.
If using \code{incidence_period = "epiweek"}, the response should
be something for which summing daily values over an epiweek makes sense
(e.g., counts or proportions but not log(counts) or log(proportions)).
Available data sources and signals are documented in the \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html}{COVIDcast signal documentation}.
A few optional columns are also allowed. If not specified, these will default
to the values of the similarly named argument.

A column \code{start_day} can be included. This can be a \link{Date}
object or string in the form "YYYY-MM-DD", indicating the earliest date of
data needed from that data source. Importantly, \code{start_day} can also be a
function (represented as a list column) that takes a forecast date and
returns a start date for model training (again, Date object or string in
the form "YYYY-MM-DD"). The latter is useful when the start date should be
computed dynamically from the forecast date (e.g., when \code{forecaster} only
trains on the most recent 4 weeks of data).

You may also include a \code{geo_type} column, a \code{geo_values} column and/or an
\code{as_of} column.
The first two should contain a string. If unspecified, these will have
the same defaults as \code{covidcast::covidcast_signal()}, namely
\code{geo_type = "county"} and \code{geo_values = "*"}.
These arguments allow you to download different data than
what you're actually trying to predict, say using state-level data to
predict national outcomes.

By default, the \code{as_of} date of data downloaded from
COVIDcast is loaded with \code{as_of = forecast_date}. This means that data
is "rewound" to days in the past. Any data revisions made since, would
not have been present at that time, and would not be available to the
forecaster. It's likely, for example, that no data would actually exist
for the forecast date on the forecast date (there is some latency between
the time signals are reported and the dates for which they are reported).
You can override this functionality, though we strongly advise you do so
with care, by passing a function of the forecast_date or a single date
here. The function should return a \link{Date}.}

\item{forecast_dates}{Vector of Date objects (or strings of the form
"YYYY-MM-DD") indicating dates on which forecasts will be made.}

\item{incidence_period}{String indicating the incidence period, either
"epiweek" or "day".}

\item{apply_corrections}{an optional function that applies data corrections
to the signals. Input is a data frame or list as returned as
\code{df <- covidcast::download_signals()}.
The returned object should be of the type expected by your forecaster.
This function will be called as \code{apply_corrections(df)}.}

\item{response_data_source}{String indicating the \code{data_source} of the response.
This is used mainly for downstream evaluation. By default, this will be the
same as the \code{data_source} in the first row of the \code{signals} tibble.}

\item{response_data_signal}{String indicating the \code{signal} of the response.
This is used mainly for downstream evaluation. By default, this will be the
same as the \code{signal} in the first row in the \code{signals} tibble.}

\item{forecaster_args}{a list of additional named arguments to be passed
to \code{forecaster()}. A common use case would be to pass the period ahead
(e.g. predict 1 day, 2 days, ..., k days ahead). Note that \code{ahead} is a
required component of the forecaster output (see above).}

\item{parallel_execution}{TRUE, FALSE, or a single positive integer. If TRUE,
uses bettermc::mclapply to execute each forecast date prediction in
parallel on max number of cores - 1. If FALSE, the code is run on a single
core. If integer, runs in parallel on that many cores, clipping to the max
number of detected cores if a greater number is requested.}

\item{additional_mclapply_args}{a named list of additional arguments to pass
to \code{\link[bettermc:mclapply]{bettermc::mclapply}} (besides \code{X}, \code{FUN}, and
\code{mc.cores}.)}

\item{honest_as_of}{a boolean that, if true, ensures that the forecast_day, end_day,
and as_of are all equal when downloading data. Otherwise, as_of is allowed to be freely
set (and defaults to current date if not presented).}

\item{offline_signal_dir}{the directory that stores the cached data for each
(signal, forecast day) pair. If this is null, no caching is done and the data is
downloaded from covidcast.}
}
\value{
Long data frame of forecasts with a class of \code{predictions_cards}.
The first 4 columns are the same as those returned by the forecaster. The
remainder specify the prediction task, 10 columns in total:
\code{ahead}, \code{geo_value}, \code{quantile}, \code{value}, \code{forecaster}, \code{forecast_date},
\code{data_source}, \code{signal}, \code{target_end_date}, and \code{incidence_period}. Here
\code{data_source} and \code{signal} correspond to the response variable only.
}
\description{
For each of the provided forecast dates, runs a forecaster using the data
that would have been available as of that given forecast date. Returns a list
of "predictions cards", where each list element corresponds to a different
forecast date. A predictions card is a data frame giving the forecast
distributions of a given forecaster for a given forecast task. A forecast
task is specified by the forecast date, ahead, response, incidence period,
and geo type (e.g., 1-epiweek-ahead death forecasting at the state level with
predictions made using the information as of September 14).
}
\examples{
\dontrun{
baby_predictions = get_predictions(
  baseline_forecaster, "baby",
  tibble::tibble(
    data_source = "jhu-csse",
    signal = "deaths_incidence_num",
    start_day = "2020-08-15",
    geo_values = "mi",
    geo_type = "state"),
  forecast_dates = "2020-10-01",
  incidence_period = "epiweek",
  forecaster_args = list(
    incidence_period = "epiweek",
    ahead = 1:4
  ))
}

}
