% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions_single_date}
\alias{get_predictions_single_date}
\title{Get predictions cards for a single date}
\usage{
get_predictions_single_date(
  forecaster,
  name_of_forecaster,
  signals,
  forecast_date,
  incidence_period,
  ahead,
  geo_type,
  geo_values,
  ...
)
}
\description{
Get predictions cards for a single date
}
