% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Plot calibration curves}
\usage{
plot_calibration(scorecard, type = c("wedgeplot", "traditional"), alpha = 0.1)
}
\arguments{
\item{scorecard}{Single score card.}

\item{type}{One of "wedgeplot" or "traditional".}

\item{alpha}{Parameter defining nominal interval coverage.}
}
\description{
Plot calibration curves
}
