% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Plot calibration curves}
\usage{
plot_calibration(
  scorecard,
  type = c("wedgeplot", "traditional"),
  grp_vars = c("forecaster", "forecast_date", "ahead"),
  avg_vars = c("geo_value"),
  legend_position = "bottom"
)
}
\arguments{
\item{scorecard}{Single score card.}

\item{type}{One of "wedgeplot" or "traditional".}

\item{grp_vars}{}

\item{avg_vars}{}

\item{legend_position}{}
}
\description{
Plot calibration curves
}
