% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Plot calibration curves}
\usage{
plot_calibration(
  scorecard,
  type = c("wedgeplot", "traditional"),
  grp_vars = c("forecaster", "forecast_date", "ahead"),
  avg_vars = c("location"),
  legend_position = "bottom"
)
}
\arguments{
\item{scorecard}{Single score card.}

\item{type}{One of "wedgeplot" or "traditional".}

\item{legend_position}{Legend position, the default being "bottom".}

\item{alpha}{Deprecated parameter to be removed soon.}
}
\description{
Plot calibration curves
}
