% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_canonical.R
\name{plot_canonical}
\alias{plot_canonical}
\title{A canonical function for creating plots within evalcast}
\usage{
plot_canonical(
  df,
  x,
  y,
  aggr = mean,
  dots = TRUE,
  lines = TRUE,
  group_vars = "forecaster",
  facet_rows = NULL,
  facet_cols = NULL,
  base_forecaster = NULL,
  scale_before_aggr = FALSE,
  title = waiver(),
  subtitle = waiver(),
  xlab = waiver(),
  ylab = waiver(),
  legend_position = "bottom",
  legend_title = NULL
)
}
\arguments{
\item{df}{dataframe containing the data to plot}

\item{x}{Name of the variable in \code{df} to be used for the x-axis}

\item{y}{name of the variable in \code{df} to be used for the y-axis}

\item{aggr}{function to use to aggregate data over \code{group_vars}}

\item{dots}{when set to TRUE, a dot layer is included in the plot}

\item{lines}{when set to TRUE, a line layer is included in the plot}

\item{group_vars}{The set of variable names in \code{df} which to aggregate over.
"forecaster" by default.}

\item{facet_rows}{A variable name to facet data over. Creates a
separate row of plots for each value of specified variable. Can be used
with \code{facet_cols} to create a grid of plots.}

\item{facet_cols}{Same as \code{facet_rows}, but with columns.}

\item{base_forecaster}{If set, scales they y-value for all forecasters by
the corresponding values of the base_forecaster}

\item{title}{title of the plot, if set}

\item{subtitle}{subtitle of the plot, if set}

\item{xlab}{x-axis label, if set}

\item{ylab}{y-axis label, if set}

\item{legend_position}{where to position the legend ("bottom" by default)}

\item{legend_title}{title of the legend, if set}

\item{scale_before_aggregation}{If TRUE, scales results by the
\code{base_forecaster} before aggregating by the \code{aggr} function}
}
\description{
A canonical function for creating plots within evalcast
}
