% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Produce trajectory plots of forecasts and observed data}
\usage{
plot_trajectory(
  list_of_predictions_cards,
  first_day = "2020-07-01",
  last_day = Sys.Date(),
  alpha = 0.2,
  nrow = 6,
  ncol = 4,
  page_facet = 1,
  cutoff = 52
)
}
\arguments{
\item{list_of_predictions_cards}{collection of prediction cards as returned
by \code{get_predictions()}. Supports collections of multiple forecast dates and
forecasters}

\item{first_day}{the earliest date to display}

\item{last_day}{the latest date to display. Defaults to the most recent data
available (plus future, if forecasts are past today)}

\item{alpha}{Displayed ribbons are the 1-alpha coverage results}
}
\value{
Produces a ggplot object
}
\description{
Produce trajectory plots of forecasts and observed data
}
\examples{
sister_preds = get_predictions(
  baseline_forecaster, "sister",
  tibble::tibble(data_source="usa-facts", signal = "deaths_incidence_num",
    start_day=lubridate::ymd("2020-07-01")),
    lubridate::ymd("2020-09-01"),"epiweek", 1:4, "state", c("mi"))
baby_preds = get_predictions(
  baseline_forecaster, "baby",
  tibble::tibble(data_source="usa-facts", signal = "deaths_incidence_num",
    start_day=lubridate::ymd("2020-07-01")),
    lubridate::ymd("2020-09-08"),"epiweek", 1:4, "state", c("mi"))
 plot_trajectory(c(sister_preds, baby_preds), last_day="2020-10-01")
}
