% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{print.prediction_card}
\alias{print.prediction_card}
\title{Print a single prediction card.}
\usage{
\method{print}{prediction_card}(card, ...)
}
\arguments{
\item{card}{Prediction card.}
}
\description{
Print a single prediction card.
}
