% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{scale_by_forecaster}
\alias{scale_by_forecaster}
\title{Scale error measures based on those of a particular forecaster.}
\usage{
scale_by_forecaster(
  score_card,
  vars,
  denom_forecaster,
  err_cols = c("ae", "wis")
)
}
\arguments{
\item{score_card}{score_card like that returned by \code{evaluate_predictions()}}

\item{vars}{vector of column names in \code{score_card} to normalize}

\item{denom_forecaster}{name of forecaster in \code{score_card$forecaster} column by whose error
values the remaining forecasters' errors will be scaled}

\item{err_cols}{vector of column names in \code{score_card} that contain error measures.
Elements of \code{err_cols} that are not in \code{vars} will be dropped from the final output to
avoid a mix of scaled and unscaled measures.}
}
\value{

}
\description{
Scale error measures based on those of a particular forecaster.
}
