% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{unique_attr}
\alias{unique_attr}
\title{Return unique value of attribute or throw error}
\usage{
unique_attr(cards, attribute)
}
\arguments{
\item{cards}{a list of predictions cards or a list of scorecards}

\item{attribute}{name of attribute}
}
\description{
If TRUE, returns the unique value; if FALSE, throws an error.
}
