% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_cards.R
\name{unpack_single_card}
\alias{unpack_single_card}
\title{Unpack a single prediction or evaluation card into an unnested tibble}
\usage{
unpack_single_card(card)
}
\arguments{
\item{card}{Evaluation or prediction card.}
}
\value{
See \code{aggregate_cards}.
}
\description{
This is a generic method for dispatching to specific calls for evaluation and prediction cards.
}
