% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_measures.R
\name{weighted_interval_score}
\alias{weighted_interval_score}
\title{Compute weighted interval score}
\usage{
weighted_interval_score(quantile_forecasts, actual_value)
}
\arguments{
\item{quantile_forecasts}{Tibble of quantile forecasts.}

\item{actual_value}{Actual value.}
}
\description{
Computes weighted interval score (WIS), a well-known quantile-based
approximation of the commonly-used continuous ranked probability score
(CRPS). WIS is a proper score, and can be thought of as a distributional
generalization of absolute error. For example, see \href{https://arxiv.org/abs/2005.12881}{Bracher et al. (2020)} for discussion in the context
of COVID-19 forecasting.
}
